rename data_contract_factory module
